MongoDB 実装の統合
