+++ app/api/models/takos/account.ts
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const accountSchema = new mongoose.Schema({
   userName: { type: String, required: true },
@@ -28,7 +29,8 @@ const accountSchema = new mongoose.Schema({
   },
 });
 
-accountSchema.index({ userName: 1 }, { unique: true });
+accountSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
+accountSchema.index({ userName: 1, tenant_id: 1 }, { unique: true });
 
 const Account = mongoose.models.Account ??
   mongoose.model("Account", accountSchema);
