+++ app/api/models/takos/attachment.ts
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const attachmentSchema = new mongoose.Schema({
   _id: { type: String },
@@ -9,6 +10,8 @@ const attachmentSchema = new mongoose.Schema({
   updated_at: { type: Date, default: Date.now },
   deleted_at: { type: Date },
 });
+
+attachmentSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
 
 const Attachment = mongoose.models.Attachment ??
   mongoose.model("Attachment", attachmentSchema, "attachments");
