+++ app/api/models/takos/fcm_token.ts
@@ -1,11 +1,13 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const fcmTokenSchema = new mongoose.Schema({
   token: { type: String, required: true },
   userName: { type: String, default: "" },
 });
 
-fcmTokenSchema.index({ token: 1 }, { unique: true });
+fcmTokenSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
+fcmTokenSchema.index({ token: 1, tenant_id: 1 }, { unique: true });
 
 const FcmToken = mongoose.models.FcmToken ??
   mongoose.model("FcmToken", fcmTokenSchema);
