+++ app/api/models/takos/key_package.ts
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const keyPackageSchema = new mongoose.Schema({
   userName: { type: String, required: true, index: true },
@@ -11,6 +12,9 @@ const keyPackageSchema = new mongoose.Schema({
   tenant_id: { type: String, index: true },
   createdAt: { type: Date, default: Date.now },
 });
+
+keyPackageSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
+keyPackageSchema.index({ userName: 1, tenant_id: 1 });
 
 const KeyPackage = mongoose.models.KeyPackage ??
   mongoose.model("KeyPackage", keyPackageSchema);
