+++ app/api/models/takos/message.ts
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const messageSchema = new mongoose.Schema({
   _id: { type: String },
@@ -15,6 +16,8 @@ const messageSchema = new mongoose.Schema({
     cc: { type: [String], default: [] },
   },
 });
+
+messageSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
 
 const Message = mongoose.models.Message ??
   mongoose.model("Message", messageSchema, "messages");
