+++ app/api/models/takos/notification.ts
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import tenantScope from "../plugins/tenant_scope.ts";
 
 const notificationSchema = new mongoose.Schema({
   title: { type: String, required: true },
@@ -7,6 +8,8 @@ const notificationSchema = new mongoose.Schema({
   read: { type: Boolean, default: false },
   createdAt: { type: Date, default: Date.now },
 });
+
+notificationSchema.plugin(tenantScope, { envKey: "ACTIVITYPUB_DOMAIN" });
 
 const Notification = mongoose.models.Notification ??
   mongoose.model("Notification", notificationSchema);
