+++ app/api/models/takos_host/encrypted_keypair.ts
@@ -1,16 +1,12 @@
 import mongoose from "mongoose";
-
-const encryptedKeyPairSchema = new mongoose.Schema({
-  userName: { type: String, required: true },
-  content: { type: String, required: true },
-  tenant_id: { type: String, index: true },
-  createdAt: { type: Date, default: Date.now },
-});
-
-encryptedKeyPairSchema.index({ userName: 1, tenant_id: 1 }, { unique: true });
+import { encryptedKeyPairSchema } from "../takos/encrypted_keypair.ts";
 
 const HostEncryptedKeyPair = mongoose.models.HostEncryptedKeyPair ??
-  mongoose.model("HostEncryptedKeyPair", encryptedKeyPairSchema);
+  mongoose.model(
+    "HostEncryptedKeyPair",
+    encryptedKeyPairSchema,
+    "encryptedkeypairs",
+  );
 
 export default HostEncryptedKeyPair;
 export { encryptedKeyPairSchema };
