+++ app/api/models/takos_host/encrypted_message.ts
@@ -1,18 +1,12 @@
 import mongoose from "mongoose";
-
-const encryptedMessageSchema = new mongoose.Schema({
-  roomId: { type: String, index: true },
-  from: { type: String, required: true },
-  to: { type: [String], required: true },
-  content: { type: String, required: true },
-  mediaType: { type: String, default: "message/mls" },
-  encoding: { type: String, default: "base64" },
-  tenant_id: { type: String, index: true },
-  createdAt: { type: Date, default: Date.now },
-});
+import { encryptedMessageSchema } from "../takos/encrypted_message.ts";
 
 const HostEncryptedMessage = mongoose.models.HostEncryptedMessage ??
-  mongoose.model("HostEncryptedMessage", encryptedMessageSchema);
+  mongoose.model(
+    "HostEncryptedMessage",
+    encryptedMessageSchema,
+    "encryptedmessages",
+  );
 
 export default HostEncryptedMessage;
 export { encryptedMessageSchema };
