+++ app/api/models/takos_host/fcm_token.ts
@@ -1,15 +1,8 @@
 import mongoose from "mongoose";
-
-const hostFcmTokenSchema = new mongoose.Schema({
-  token: { type: String, required: true },
-  userName: { type: String, default: "" },
-  tenant_id: { type: String, index: true },
-});
-
-hostFcmTokenSchema.index({ token: 1, tenant_id: 1 }, { unique: true });
+import { fcmTokenSchema } from "../takos/fcm_token.ts";
 
 const HostFcmToken = mongoose.models.HostFcmToken ??
-  mongoose.model("HostFcmToken", hostFcmTokenSchema);
+  mongoose.model("HostFcmToken", fcmTokenSchema, "fcmtokens");
 
 export default HostFcmToken;
-export { hostFcmTokenSchema };
+export { fcmTokenSchema };
