+++ app/api/models/takos_host/follow_edge.ts
@@ -1,16 +1,8 @@
 import mongoose from "mongoose";
-
-const followEdgeSchema = new mongoose.Schema({
-  tenant_id: { type: String, required: true },
-  actor_id: { type: String, required: true },
-  since: { type: Date, default: Date.now },
-  relay: { type: String, default: null },
-});
-
-followEdgeSchema.index({ actor_id: 1, tenant_id: 1 });
+import { followEdgeSchema } from "../takos/follow_edge.ts";
 
 const HostFollowEdge = mongoose.models.HostFollowEdge ??
-  mongoose.model("HostFollowEdge", followEdgeSchema);
+  mongoose.model("HostFollowEdge", followEdgeSchema, "follow_edge");
 
 export default HostFollowEdge;
 export { followEdgeSchema };
