+++ app/api/models/takos_host/notification.ts
@@ -1,16 +1,8 @@
 import mongoose from "mongoose";
-
-const notificationSchema = new mongoose.Schema({
-  title: { type: String, required: true },
-  message: { type: String, required: true },
-  type: { type: String, default: "info" },
-  read: { type: Boolean, default: false },
-  tenant_id: { type: String, index: true },
-  createdAt: { type: Date, default: Date.now },
-});
+import { notificationSchema } from "../takos/notification.ts";
 
 const HostNotification = mongoose.models.HostNotification ??
-  mongoose.model("HostNotification", notificationSchema);
+  mongoose.model("HostNotification", notificationSchema, "notifications");
 
 export default HostNotification;
 export { notificationSchema };
