+++ app/api/models/takos_host/session.ts
@@ -1,14 +1,8 @@
 import mongoose from "mongoose";
-
-const sessionSchema = new mongoose.Schema({
-  sessionId: { type: String, required: true, unique: true },
-  expiresAt: { type: Date, required: true, index: { expires: 0 } },
-  createdAt: { type: Date, default: Date.now },
-  tenant_id: { type: String, index: true },
-});
+import { sessionSchema } from "../takos/session.ts";
 
 const HostSession = mongoose.models.HostSession ??
-  mongoose.model("HostSession", sessionSchema);
+  mongoose.model("HostSession", sessionSchema, "sessions");
 
 export default HostSession;
 export { sessionSchema };
